requirements.txt
openai~=1.68.2
fastapi~=0.115.12
python-dotenv~=1.1.0